Refactor Settings data object structure and update event triggering
The structure of Settings has been moved into a separate package (`data.settings`) for event addons coming in the next commit.
