+++ src/main/java/online/hatsunemiku/tachideskvaadinui/services/SettingsService.java
@@ -11,8 +11,7 @@ import java.nio.file.Path;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
-import online.hatsunemiku.tachideskvaadinui.data.Settings;
-import online.hatsunemiku.tachideskvaadinui.startup.TachideskMaintainer;
+import online.hatsunemiku.tachideskvaadinui.data.settings.Settings;
 import online.hatsunemiku.tachideskvaadinui.utils.PathUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
