+++ src/main/java/online/hatsunemiku/tachideskvaadinui/startup/TachideskMaintainer.java
@@ -13,6 +13,7 @@ import java.util.Optional;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import online.hatsunemiku.tachideskvaadinui.data.Meta;
+import online.hatsunemiku.tachideskvaadinui.data.settings.Settings;
 import online.hatsunemiku.tachideskvaadinui.startup.download.ReadableProgressByteChannel;
 import online.hatsunemiku.tachideskvaadinui.utils.PathUtils;
 import online.hatsunemiku.tachideskvaadinui.utils.SerializationUtils;
