+++ src/main/java/online/hatsunemiku/tachideskvaadinui/utils/MangaDataUtils.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 import java.util.List;
 import lombok.experimental.UtilityClass;
 import online.hatsunemiku.tachideskvaadinui.data.tachidesk.Chapter;
-import online.hatsunemiku.tachideskvaadinui.data.Settings;
+import online.hatsunemiku.tachideskvaadinui.data.settings.Settings;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.web.client.RestTemplate;
