+++ src/main/java/online/hatsunemiku/tachideskvaadinui/view/SearchView.java
@@ -23,7 +23,7 @@ import lombok.extern.slf4j.Slf4j;
 import online.hatsunemiku.tachideskvaadinui.component.card.MangaCard;
 import online.hatsunemiku.tachideskvaadinui.component.combo.LangComboBox;
 import online.hatsunemiku.tachideskvaadinui.component.events.source.LanguageListChangeEvent;
-import online.hatsunemiku.tachideskvaadinui.data.Settings;
+import online.hatsunemiku.tachideskvaadinui.data.settings.Settings;
 import online.hatsunemiku.tachideskvaadinui.data.tachidesk.Manga;
 import online.hatsunemiku.tachideskvaadinui.data.tachidesk.Source;
 import online.hatsunemiku.tachideskvaadinui.data.tachidesk.search.SearchResponse;
